backend/requirements.txt
flask[async]==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
ollama==0.1.6
python-magic==0.4.27
python-magic-bin==0.4.27